Change file structure in plugin (#5636)
Requires #5599.

## Summary

Currently files in plugin were named after their most important
function. This was a temporary solution applied in the past. Now files
are properly aggregated by their usage domain.

## Test plan

All tests and CI pass.
