+++ plugin/src/workletFactory.ts
@@ -39,7 +39,7 @@ import {
 } from '@babel/types';
 import { strict as assert } from 'assert';
 import { relative } from 'path';
-import { buildWorkletString } from './buildWorkletString';
+import { buildWorkletString } from './workletStringCode';
 import { globals } from './globals';
 import type { ReanimatedPluginPass, WorkletizableFunction } from './types';
 import { isRelease } from './utils';
