+++ plugin/src/workletFactoryCall.ts
@@ -2,7 +2,7 @@ import type { NodePath } from '@babel/core';
 import { callExpression } from '@babel/types';
 import type { CallExpression } from '@babel/types';
 import type { ReanimatedPluginPass, WorkletizableFunction } from './types';
-import { makeWorkletFactory } from './makeWorklet';
+import { makeWorkletFactory } from './workletFactory';
 
 export function makeWorkletFactoryCall(
   path: NodePath<WorkletizableFunction>,
