+++ plugin/src/workletStringCode.ts
